advance_ai_agents/trend_analyzer_agent/requirements.txt
google-adk
exa-py
tavily-python
firecrawl-py
python-dotenv 